requirements.txt
google-api-python-client
gspread
oauth2client
python-telegram-bot
python-dotenv
requests
beautifulsoup4
pytest
